feature/PTS-847-create-traits - created traits
