+++ packages/EasyErrorHandler/src/Exceptions/Traits/SubCodeAwareExceptionTrait.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyErrorHandler\Exceptions\Traits;
+
+trait SubCodeAwareExceptionTrait
+{
+    /**
+     * @var int
+     */
+    protected $subCode = 0;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSubCode(): int
+    {
+        return $this->subCode;
+    }
+
+    /**
+     * Sets the sub code for an exception.
+     */
+    public function setSubCode(int $subCode): self
+    {
+        $this->subCode = $subCode;
+
+        return $this;
+    }
+}
