+++ packages/EasyErrorHandler/src/Exceptions/Traits/ValidationExceptionTrait.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyErrorHandler\Exceptions\Traits;
+
+trait ValidationExceptionTrait
+{
+    /**
+     * @var mixed[]
+     */
+    protected $errors = [];
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getErrors(): array
+    {
+        return $this->errors;
+    }
+
+    /**
+     * Sets validation errors.
+     *
+     * @param mixed[] $errors
+     */
+    public function setErrors(array $errors): self
+    {
+        $this->errors = $errors;
+
+        return $this;
+    }
+}
