+++ packages/EasyErrorHandler/src/Exceptions/ValidationException.php
@@ -4,32 +4,10 @@ declare(strict_types=1);
 
 namespace EonX\EasyErrorHandler\Exceptions;
 
+use EonX\EasyErrorHandler\Exceptions\Traits\ValidationExceptionTrait;
 use EonX\EasyErrorHandler\Interfaces\ValidationExceptionInterface;
 
 abstract class ValidationException extends BadRequestException implements ValidationExceptionInterface
 {
-    /**
-     * @var mixed[]
-     */
-    protected $errors = [];
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getErrors(): array
-    {
-        return $this->errors;
-    }
-
-    /**
-     * Sets validation errors.
-     *
-     * @param mixed[] $errors
-     */
-    public function setErrors(array $errors): self
-    {
-        $this->errors = $errors;
-
-        return $this;
-    }
+    use ValidationExceptionTrait;
 }
